keep the worker module inside the cap: scope planning moves to the run (issue #384)
